feat(granite): refactor buffer allocator for all kinds of buffers
All the get*Pointer methods now return BufferSubAllocator, it allows
the caller to perform multiple sub-allocations and also allows multiple
buffers to be suballocated in parallel (for UBO in future).

It also allows specifying an extra alignment offset for the start of
each new vertex format, so that buffer binding offsets can be expressed
using baseInstance and baseVertex, reducing the number of buffer bindings.

For mesh draw, we use the concepts of append and static data to simplify.
Improvements to UBO allocation, exception handling, and further code
cleanup will be done later CI. And SSBO is reserved for gradient data.