feat(granite): improve paint color handling and fix
If step solid color is used, the color of KeyContext will no longer be calculated since it's not used.
If shader is constant color, reduce to solid color.
Fixed the case where step solid color is not used and the paint is solid color,
then it should be in destination space.